Updated the example project to use the latest version
