--- example/.watchmanconfig
@@ -1,1 +1,0 @@
-{}
